[PM] Remove a failed attempt to port the CallGraph analysis to the new pass manager.
This never worked, and won't ever work. It was actually why I ended up
building the LazyCallGraph set of code which is more more effectively
wired up to the new pass manager. This accidentally got committed when
I was trying to land a cleanup of the code organization in the other
parts of this file. =[ My bad, but fortunately Dave was keen eyed enough
to spot that this code couldn't possibly work. =]

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@244127 91177308-0d34-0410-b5e6-96231b3b80d8