[MC] Split the layout part of MCAssembler::finish() into its own method. NFC.
Split a MCAssembler::layout() method out of MCAssembler::finish(). This allows
running the MCSections layout separately from the streaming of the output
file. This way if a client wants to use MC to generate section contents, but
emit something different than the standard relocatable object files it is
possible (llvm-dsymutil is such a client).

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@246008 91177308-0d34-0410-b5e6-96231b3b80d8